[REFACTOR] 인증 외부 클라이언트 패키지 분리
