+++ src/main/java/org/websoso/WSSServer/application/AccountApplication.java
@@ -10,11 +10,9 @@ import org.websoso.WSSServer.feed.feed.service.FeedServiceImpl;
 import org.websoso.WSSServer.infrastructure.discord.DiscordMessageClient;
 import org.websoso.WSSServer.infrastructure.discord.DiscordWebhookMessage;
 import org.websoso.WSSServer.dto.user.WithdrawalRequest;
-import org.websoso.WSSServer.feed.comment.application.CommentManagementApplication;
-import org.websoso.WSSServer.feed.feed.application.FeedManagementApplication;
 import org.websoso.WSSServer.auth.repository.RefreshTokenRepository;
 import org.websoso.WSSServer.auth.service.AppleService;
-import org.websoso.WSSServer.auth.service.KakaoService;
+import org.websoso.WSSServer.auth.client.KakaoService;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.user.domain.WithdrawalReason;
 import org.websoso.WSSServer.user.message.UserDiscordMessageFormatter;
