+++ src/main/java/org/websoso/WSSServer/auth/application/AuthApplication.java
@@ -8,24 +8,23 @@ import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.websoso.WSSServer.auth.controller.dto.ApplePublicKeys;
-import org.websoso.WSSServer.auth.controller.dto.AppleTokenResponse;
+import org.websoso.WSSServer.auth.client.dto.AppleTokenResponse;
 import org.websoso.WSSServer.auth.controller.dto.AuthResponse;
-import org.websoso.WSSServer.auth.controller.dto.KakaoUserInfo;
 import org.websoso.WSSServer.auth.controller.dto.LogoutRequest;
 import org.websoso.WSSServer.auth.controller.dto.ReissueResponse;
 import org.websoso.WSSServer.auth.domain.RefreshToken;
-import org.websoso.WSSServer.auth.domain.UserAppleToken;
 import org.websoso.WSSServer.auth.jwt.CustomAuthenticationToken;
 import org.websoso.WSSServer.auth.jwt.JWTUtil;
 import org.websoso.WSSServer.auth.jwt.JwtProvider;
 import org.websoso.WSSServer.auth.jwt.JwtValidationType;
 import org.websoso.WSSServer.auth.repository.RefreshTokenRepository;
-import org.websoso.WSSServer.auth.service.AppleClient;
-import org.websoso.WSSServer.auth.service.AppleKeyGenerator;
+import org.websoso.WSSServer.auth.client.AppleClient;
+import org.websoso.WSSServer.auth.client.AppleKeyGenerator;
 import org.websoso.WSSServer.auth.service.AppleService;
-import org.websoso.WSSServer.auth.service.KakaoService;
+import org.websoso.WSSServer.auth.client.KakaoService;
 import org.websoso.WSSServer.auth.service.TokenService;
+import org.websoso.WSSServer.auth.client.dto.ApplePublicKeys;
+import org.websoso.WSSServer.auth.client.dto.KakaoUserInfo;
 import org.websoso.WSSServer.dto.user.LoginResponse;
 import org.websoso.WSSServer.exception.exception.CustomAuthException;
 import org.websoso.WSSServer.user.domain.User;
