+++ src/main/java/org/websoso/WSSServer/auth/service/AppleService.java
@@ -38,10 +38,8 @@ import java.util.Base64;
 import java.util.Date;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.bouncycastle.util.io.pem.PemObject;
 import org.bouncycastle.util.io.pem.PemReader;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Service;
@@ -50,17 +48,13 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
 import org.websoso.WSSServer.auth.controller.dto.AppleIdUpdateRequest;
-import org.websoso.WSSServer.auth.controller.dto.AppleLoginRequest;
-import org.websoso.WSSServer.auth.controller.dto.ApplePublicKey;
-import org.websoso.WSSServer.auth.controller.dto.ApplePublicKeys;
-import org.websoso.WSSServer.auth.controller.dto.AppleTokenResponse;
-import org.websoso.WSSServer.auth.controller.dto.AuthResponse;
-import org.websoso.WSSServer.auth.domain.RefreshToken;
+import org.websoso.WSSServer.auth.client.dto.AppleTokenResponse;
 import org.websoso.WSSServer.auth.domain.UserAppleToken;
-import org.websoso.WSSServer.auth.jwt.CustomAuthenticationToken;
 import org.websoso.WSSServer.auth.jwt.JwtProvider;
 import org.websoso.WSSServer.auth.repository.RefreshTokenRepository;
 import org.websoso.WSSServer.auth.repository.UserAppleTokenRepository;
+import org.websoso.WSSServer.auth.client.dto.ApplePublicKey;
+import org.websoso.WSSServer.auth.client.dto.ApplePublicKeys;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.exception.exception.CustomAppleLoginException;
 import org.websoso.WSSServer.user.repository.UserRepository;
